Single class per file
