+++ src/JTracker/Controller/AbstractAjaxController.php
@@ -100,45 +100,3 @@ abstract class AbstractAjaxController extends AbstractTrackerController
 	 */
 	abstract protected function prepareResponse();
 }
-
-/**
- * AJAX response object
- *
- * @since  1.0
- */
-class AjaxResponse
-{
-	/**
-	 * Data object.
-	 *
-	 * @var    \stdClass
-	 * @since  1.0
-	 */
-	public $data;
-
-	/**
-	 * Error message.
-	 *
-	 * @var    string
-	 * @since  1.0
-	 */
-	public $error = '';
-
-	/**
-	 * Message string.
-	 *
-	 * @var    string
-	 * @since  1.0
-	 */
-	public $message = '';
-
-	/**
-	 * Constructor
-	 *
-	 * @since  1.0
-	 */
-	public function __construct()
-	{
-		$this->data = new \stdClass;
-	}
-}
